Bubbles inhibits Spykes attack increase
In a team with Bubbles + Spyke, spykes +10 each round is decreased by bubbles skill. This is most definitly not wanted, and strange.
It could be changed to use any enemys bubbles skill, however if it was Bubbles + Putrid this would be bad aswell.
So pull request to remove bubbles skill effecting iAtk.