Sanqueen is assymetric
Sanqueen code checks that she is alive before applying leech. However the check is sort of wrong, as the two sides takes turn to fight. So the to go will not see her as dead (yet) and apply heal. That heal will go to the next unit. However the second to go will see her as dead and not apply the heal to next unit. 

I dont see an easy solution with current code. 
1) Either remove the check, so she always heal next unit (strange, but fair).
2) Stop removing dead units in subturns.